requirements.txt
BeautifulSoup==3.2.1
Django==1.4.3
Fabric==1.5.1
South==0.7.6
argparse==1.2.1
cmsplugin-filer==0.9.3
cmsplugin-zinnia==0.3
distribute==0.6.24
django-blog-zinnia==0.12
django-classy-tags==0.3.4.1
django-cms==2.3.5
django-cms-smartsnippets==0.1.12
django-filer==0.9.3
django-localflavor-us==1.0
django-mptt==0.5.2
django-paypal==0.1.2
django-polymorphic==0.2
django-reversion==1.6.5
django-sekizai==0.6.1
django-social-auth==0.7.14
django-tagging==0.3.1
django-xmlrpc==0.1.4
easy-thumbnails==1.1
feedparser==5.1.3
gunicorn==0.17.2
html5lib==0.95
httplib2==0.7.7
oauth2==1.5.211
paramiko==1.9.0
pycrypto==2.6
pyparsing==1.5.6
python-openid==2.2.5
pytz==2012j
wsgiref==0.1.2
